refactor(providers): compose transports
- Route provider streaming through shared composed transport machinery
- Split OpenAI message projection and stream processing from provider wrappers
- Preserve provider request golden behavior while removing stale dead-code suppressions
