+++ crates/aion-providers/src/lib.rs
@@ -1,15 +1,19 @@
 pub mod anthropic;
 pub mod anthropic_shared;
 pub mod bedrock;
+pub(crate) mod composed;
 pub mod error;
 pub(crate) mod framing;
 pub mod openai;
+pub(crate) mod openai_messages;
 pub(crate) mod parser;
 pub(crate) mod projector;
 pub mod provider;
 pub mod retry;
+pub(crate) mod stream_process;
 pub(crate) mod stream_runner;
 mod tool_call_sanitize;
+pub(crate) mod transport;
 pub mod vertex;
 
 pub use error::ProviderError;
